+++ tests/integration-network/src/mockchain/fpi.rs
@@ -1,6 +1,7 @@
 //! Foreign procedure invocation tests on a mock chain.
 
 mod account;
+mod blind_import;
 mod common;
 mod note;
 mod raw;